requirements.txt
pillow==11.0.0
streamlit==1.41.1
